actually make the confirmation in the installer work (must've been too tired to do this properly yesterday

+++ install.sh
@@ -1,9 +1,13 @@
-read -p "This installer will install files that are part of an fcitx5 addon into /usr/share/fcitx5/, /usr/lib/fcitx5/ and ~/.config/unixkey.json"
-git clone https://github.com/Schlafhase/UnixKey
-cd UnixKey
-cmake -S . -B build -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=Release
-cmake --build build
-sudo cmake --install build
-cp src/unixkey.json ~/.config/unixkey.json
-cd ..
-rm -rf UnixKey
+read -p "This installer will install files that are part of an fcitx5 addon into /usr/share/fcitx5/, /usr/lib/fcitx5/ and ~/.config/unixkey.json. Continue? [y/n]: " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  git clone https://github.com/Schlafhase/UnixKey
+  cd UnixKey
+  cmake -S . -B build -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=Release
+  cmake --build build
+  sudo cmake --install build
+  cp src/unixkey.json ~/.config/unixkey.json
+  cd ..
+  rm -rf UnixKey
+fi
